IT WORKS!
The event system works! Its set up to be extendable for whatever events we want. Right now i have one event that simply puts a popup their when the robot is in autonomous mode. It can probably be gotten rid of once we have actual events but it shows how to use the api.